style: Clear the warning classes that were down to a few sites each
The timestamp helpers were static in a header, so every includer got its
own unused copy. The three NotImplementedException handlers caught by
value, slicing what they caught. The switch in GetMoreTerms falls
through three times on purpose, as its comments already said, and now
says so to the compiler too.

The comparisons were each already guarded: lemma_limit is only read
after testing it is positive, mbic3_indgen_mode_ is unsigned so only its
upper bound can fail, and a negative bound is too small to count
live_count_ times just as a small positive one is.

+++ refiners/array_axiom_enumerator.cpp
@@ -358,7 +358,8 @@ bool ArrayAxiomEnumerator::check_consecutive_axioms(AxiomClass ac,
         ts_axioms_[unrolled_ax] = ax;
         num_found_lemmas++;
 
-        if (lemma_limit > 0 && num_found_lemmas >= lemma_limit) {
+        if (lemma_limit > 0
+            && num_found_lemmas >= static_cast<size_t>(lemma_limit)) {
           // if given a lemma limit, then finish when that limit is reached
           return num_found_lemmas;
         }
@@ -434,7 +435,8 @@ bool ArrayAxiomEnumerator::check_nonconsecutive_axioms(AxiomClass ac,
         to_axiom_inst_.insert({ unrolled_ax, ax_inst });
         num_found_lemmas++;
 
-        if (lemma_limit > 0 && num_found_lemmas >= lemma_limit) {
+        if (lemma_limit > 0
+            && num_found_lemmas >= static_cast<size_t>(lemma_limit)) {
           // if given a lemma limit, then finish when that limit is reached
           return num_found_lemmas;
         }
